backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-genai>=0.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0
protobuf>=3.20.0
scikit-learn>=1.3.0
tqdm>=4.65.0
